[INFO] change from restrict app to restrict games
